adds code refactoring
breaking down of main file to multiple sub files based on functions.
