+++ app/commands.py
@@ -1,0 +1,7 @@
+class Commands:
+    DECODE = "decode"
+    INFO = "info"
+    PEERS = "peers"
+    HANDSHAKE = "handshake"
+    DOWNLOAD_PIECE = "download_piece"
+    DOWNLOAD = "download"
