+++ app/message_ids.py
@@ -1,0 +1,6 @@
+class MessageIDs:
+    UNCHOKE = 1
+    INTERESTED = 2
+    BITFIELD = 5
+    REQUEST = 6
+    PIECE = 7
